requirements.txt
python-docx
openai
pyperclip